Refactor pkg/cloud/metadata.go into pkg/cloud/metadata_*.go files
